% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{left_right_xylo}
\alias{left_right_xylo}
\title{Left-right xyloplots}
\usage{
left_right_xylo(lhs, rhs, ...)
}
\arguments{
\item{lhs}{A set of values to use as the left-hand side histograms (see \code{\link{xyloplot}} for details on admissible types)..}

\item{rhs}{As \code{lhs} for the right-hand side.}

\item{...}{Additional arguments passed to \code{\link{xyloplot}}.}
}
\value{
A \code{"xyloplot"} object.
}
\description{
Create a xyloplot with independent histograms on the left and right hand side of each pivot.
}
\seealso{
\code{\link{xyloplot}}
}
