% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.xyloplot}
\alias{plot.xyloplot}
\alias{print.xyloplot}
\title{Plot a xyloplot object}
\usage{
\method{plot}{xyloplot}(x, vertical = TRUE, value_lab = "Value",
  pivots_lab = "Frequency", box = TRUE, draw_empty = TRUE, ...)

\method{print}{xyloplot}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"xyloplot"}}

\item{vertical}{Logical value determining whether to plot vertical or horizontal xylophones.}

\item{value_lab}{Text to put on value axis.}

\item{pivots_lab}{Text to put on pivots axis.}

\item{box}{Logical value determining whether to draw a box around the plot.}

\item{draw_empty}{Logical value determining whether to draw `empty' boxes in the xylophones. If \code{TRUE} (default), empty boxes will appear as lines.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\value{
Renders a xyloplot.
}
\description{
Render a xyloplot object using base graphics.
}
\seealso{
\code{\link{xyloplot}}
}
