% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xylo_positions}
\alias{xylo_positions}
\title{Get x-axis positions for \code{n} xylophones}
\usage{
xylo_positions(n)
}
\arguments{
\item{n}{Number of xylophones.}
}
\description{
Get x-axis positions for \code{n} xylophones
}
