% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xyloplot.factor}
\alias{xyloplot.factor}
\title{Function for creating xyloplots from factor vectors}
\usage{
\method{xyloplot}{factor}(x, ...)
}
\arguments{
\item{x}{Character vector of values.}

\item{...}{Other arguments passed to \code{\link{xyloplot.list}}.}
}
\description{
Function for creating xyloplots from factor vectors
}
\seealso{
xyloplot.list
}

