% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xyloplot.numeric}
\alias{xyloplot.numeric}
\title{Function for creating xyloplots from numeric vectors}
\usage{
\method{xyloplot}{numeric}(x, ...)
}
\arguments{
\item{x}{Numeric vector of values.}

\item{...}{Other arguments passed to \code{\link{xyloplot.list}}.}
}
\description{
Function for creating xyloplots from numeric vectors
}
\seealso{
xyloplot.list
}

