% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xyloplot.factor}
\alias{xyloplot.factor}
\title{Function for creating xyloplots from factor vectors}
\usage{
\method{xyloplot}{factor}(x, rhs = NULL, ...)
}
\arguments{
\item{x}{Character vector of values.}

\item{rhs}{If not \code{NULL}, causes \code{xyloplot} to split xylophones and forms the distribution(s) for right hand side xylphones, complementary to the left hand side ones reflecting \code{x}.}

\item{...}{Other arguments passed to \code{\link{xyloplot.list}}.}
}
\description{
Function for creating xyloplots from factor vectors
}
\seealso{
xyloplot.list
}
