% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xyloplot}
\alias{xyloplot}
\title{Method for creating xyloplots}
\usage{
xyloplot(x, ...)
}
\arguments{
\item{x}{Numeric vector or list of numeric vectors to use for creating xyloplots.}

\item{...}{Other arguments to be passed to \code{\link{xyloplot.list}} and \code{plot}.}
}
\description{
Method for creating xyloplots
}
\examples{
xyloplot(rnorm(1000))
xyloplot(
 x=lapply(1:3, function(mean) rnorm(mean=mean, n=1000)), 
 breaks=20,
 col=rainbow(3), 
 main="title")
}
\seealso{
xyloplot.list xyloplot.numeric
}

